added testing functions/modified other functions
added testing functions to interact with the robot and select the different tests when the robot is disconnected from the usb. Fixed missing break statements in move(). Motors are reversed to move in the robot in the proper direction, this effected the encoder such that it now returns negative values when moving forwards and positive when in reverse.